Moved extensions.rs to utils
