+++ src/init/mod.rs
@@ -2,6 +2,5 @@ pub mod device;
 pub mod initialization_registry;
 pub mod instance;
 pub mod application_feature;
-pub mod extensions;
 pub mod rosella_features;
 mod utils;
