+++ src/util/mod.rs
@@ -1,2 +1,3 @@
 pub mod allocation_callbacks;
 pub mod id;
+pub mod extensions;
